reorganize mmdit
